requirements.txt
argparse==1.4.0
numpy==1.21.1
pandas==1.3.1
scipy==1.7.1
geopandas==0.9.0
fiona==1.8.20
patool==1.12
rasterstats==0.15.0
openpyxl==3.0.7
geocube==0.0.17
netCDF4==1.5.7
xarray==0.19.0
rasterio==1.2.6
rioxarray==0.6.1
Pillow==8.3.1
opencv-python==4.5.3.56
ImageHash==4.2.1
imageio==2.9.0
rich==10.7.0
